BACKLOG-17115: Use moonstone header
