+++ src/javascript/EditPanel/EditPanelHeader/ContentPath/CompositePathEntry.jsx
