+++ src/javascript/EditPanel/EditPanelHeader/ContentPath/ContentPath.container.jsx
