+++ src/javascript/EditPanel/EditPanelHeader/ContentPath/ContentPath.gql-queries.jsx
