+++ src/javascript/EditPanel/EditPanelHeader/ContentPath/ContentPath.spec.jsx
