+++ src/javascript/EditPanel/EditPanelHeader/ContentPath/index.js
