+++ src/javascript/EditPanel/EditPanelHeader/EditPanelHeader.scss
@@ -1,0 +1,58 @@
+.headerLeft {
+    display: flex;
+    flex-grow: 1;
+    align-items: center;
+
+    & > * {
+        margin-right: var(--spacing-small);
+    }
+
+    .headerTypography {
+        flex-grow: 1;
+        width: 250px;
+    }
+}
+
+.headerRight {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.headerChips {
+    & > * {
+        margin-left: var(--spacing-small);
+    }
+}
+
+.padder {
+    height: var(--spacing-medium);
+}
+
+.mainActions {
+    margin-left: var(--spacing-small);
+}
+
+.menu {
+    height: 38px;
+}
+
+.saveActions {
+    position: relative;
+
+    margin-right: var(--spacing-small);
+}
+
+.publishActions {
+    position: relative;
+
+    & button:not(:first-child) {
+        margin-left: 1px;
+    }
+}
+
+.headerToolBar {
+    display: flex;
+    align-items: center;
+    margin-top: 1px;
+}
