+++ src/javascript/EditPanel/EditPanelHeader/HeaderBadges/HeaderBadges.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import PublicationInfoBadge from '~/PublicationInfo/PublicationInfoBadge';
 import LockInfoBadge from '~/Lock/LockInfo.badge';
 import WipInfoChip from '~/EditPanel/WorkInProgress/Chip/WipInfo.Chip';
-import {UnsavedChip} from '~/EditPanel/header';
+import {UnsavedChip} from '~/EditPanel/EditPanelHeader';
 import {Constants} from '~/ContentEditor.constants';
 import styles from './HeaderBadges.scss';
