+++ src/javascript/EditPanel/EditPanelHeader/HeaderBadges/HeaderBadges.scss
