+++ src/javascript/EditPanel/EditPanelHeader/HeaderBadges/HeaderBadges.spec.jsx
