+++ src/javascript/EditPanel/EditPanelHeader/HeaderBadges/index.js
