+++ src/javascript/EditPanel/EditPanelHeader/PublishMenu.scss
@@ -1,0 +1,3 @@
+.menu {
+    height: 38px;
+}
