+++ src/javascript/EditPanel/EditPanelHeader/index.js
@@ -1,0 +1,3 @@
+export {EditPanelHeader} from './EditPanelHeader';
+export {UnsavedChip} from '~/EditPanel/EditPanelHeader/UnsavedChip';
+export {PublishMenu} from '~/EditPanel/EditPanelHeader/PublishMenu';
