--- src/javascript/EditPanel/header/ContentBreadcrumb/ContentBreadcrumb.scss
@@ -1,6 +1,0 @@
-.contentBreadcrumb {
-    display: inline-flex;
-    flex-grow: 2;
-    flex-direction: row;
-    align-items: center;
-}
