--- src/javascript/EditPanel/header/ContentBreadcrumb/ContentType/ContentType.scss
@@ -1,3 +1,0 @@
-.contentType {
-    padding: 1px 0;
-}
