--- src/javascript/EditPanel/header/ContentBreadcrumb/ContentType/index.js
@@ -1,3 +1,0 @@
-import ContentType from './ContentType';
-
-export default ContentType;
