--- src/javascript/EditPanel/header/ContentBreadcrumb/index.js
@@ -1,3 +1,0 @@
-import ContentBreadcrumb from './ContentBreadcrumb';
-
-export default ContentBreadcrumb;
