--- src/javascript/EditPanel/header/HeaderLowerSection.scss
@@ -1,5 +1,0 @@
-.headerToolBar {
-    display: flex;
-    align-items: center;
-    margin-top: 1px;
-}
