--- src/javascript/EditPanel/header/HeaderUpperSection.scss
@@ -1,57 +1,0 @@
-.header {
-    display: flex;
-    margin: var(--spacing-small) auto;
-
-    .headerLeft {
-        display: flex;
-        flex-grow: 1;
-        align-items: center;
-
-        & > * {
-            margin-right: var(--spacing-small);
-        }
-
-        .headerTypography {
-            flex-grow: 1;
-            width: 250px;
-        }
-    }
-
-    .headerRight {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-    }
-
-    .headerChips {
-        & > * {
-            margin-left: var(--spacing-small);
-        }
-    }
-}
-
-.padder {
-    height: var(--spacing-medium);
-}
-
-.mainActions {
-    margin-left: var(--spacing-small);
-}
-
-.menu {
-    height: 38px;
-}
-
-.saveActions {
-    position: relative;
-
-    margin-right: var(--spacing-small);
-}
-
-.publishActions {
-    position: relative;
-
-    & button:not(:first-child) {
-        margin-left: 1px;
-    }
-}
