--- src/javascript/EditPanel/header/index.js
@@ -1,4 +1,0 @@
-export * from './HeaderLowerSection';
-export * from './HeaderUpperSection';
-export {UnsavedChip} from '~/EditPanel/header/UnsavedChip';
-export {PublishMenu} from '~/EditPanel/header/PublishMenu';
