+++ src/javascript/utils/getButtonRenderer.jsx
@@ -6,7 +6,7 @@ import styles from './getButtonRenderer.scss';
 
 export const getButtonRenderer = ({labelStyle, defaultButtonProps, noIcon} = {}) => {
     const ButtonRenderer = props => {
-        const {addWarningBadge, buttonLabelNamespace, buttonLabelShort, buttonLabel, isVisible, buttonLabelParams, buttonIcon, actionKey, enabled, disabled, onClick, buttonProps, dataSelRole} = props;
+        const {addWarningBadge, buttonLabelNamespace, buttonLabelShort, buttonLabel, isVisible, buttonLabelParams, buttonIcon, actionKey, enabled, disabled, onClick, buttonProps, dataSelRole, className} = props;
         const {t} = useTranslation(buttonLabelNamespace);
 
         let label = buttonLabel;
@@ -27,6 +27,7 @@ export const getButtonRenderer = ({labelStyle, defaultButtonProps, noIcon} = {})
 
         let button = (
             <Button data-sel-role={dataSelRole || actionKey}
+                    className={className}
                     label={t(label, buttonLabelParams)}
                     icon={icon}
                     isDisabled={enabled === false || disabled}
@@ -62,7 +63,8 @@ export const getButtonRenderer = ({labelStyle, defaultButtonProps, noIcon} = {})
         onClick: PropTypes.func,
         buttonProps: PropTypes.object,
         addWarningBadge: PropTypes.bool,
-        dataSelRole: PropTypes.string
+        dataSelRole: PropTypes.string,
+        className: PropTypes.string
     };
 
     return ButtonRenderer;
